+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## next
  * bugfix angular.module("MyMod").bootstrap(e, [], {}) disrupting chaining
+ * API now returns {src <string>, errors <array of strings>}
 
 ## v0.8.0 2014-05-09
  * ngRoute support: $routeProvider.when("path", { .. })